moving PDB to toplevel
the firewall no longer has complete control over the protocol db

The main idea here is that the firewall was essentially acting like a
facade of the protocolDB for no reason.
The current solution is less than optimal, because really the gui
shouldn't care about the db, but there really insn't a better place for
it currently.

-------------------------------------
This also brings up the point that the zones act the same way as the pdb
did, only without being a facade... it really is part of the firewall,
but it really shouldn't be.

future endevors might include adding in a firewall manager class that
the gui talks to instead of the protocol db and firewall, which also has
the ability to "swap out" firewall instances arbitrarily (that is to
say, there would be several configureations that the manager knows
about, like "gaming" or "work" and could simply switch them at a single
click)
this consept however brings up some problems, like for instance, what if
one profile has a UDP myprotocol and the other also has myprotocol.
currently protocols are global to the program, and this has great merit,
but in this case we wouldn't want to combine the protocols, the
protocols for one profile should be distinct, but there needs to be a
way to make certian ones global. or something.
i think using the classifications might be helpful. like have the
classification but "User Defined" for global ones, and
"UDP-nameoffirewall" or something for not global ones, then also
changing the uniq id to be uniq name in class, instead of just uniq
name. that would be more combersome tho. it may be worth it in the long
run however. more thinking is required.