adding firewall manager file
the firewall manager will be the main arbiter between the user level
code and all the data. it will have the xml parser as well.
work started on adding back in the xml parser, but it was so terribly
done the first time, i am looking at seeing if there is a better way.
we'll see.