progress... *please dont try to build this...
the xml parser object builds. the ptotocol data base is mostly done,
some things may need to be added to make the program actually work, but
the firewall manager needs done first, and to do that the firewall class
needs completly changed.
the next thing is to remove the zone control things from the firewall
class, and make a seperate class to manage that.
after that changeing the firewall to not have control over data will be
needed, it will need to request infomation from the firewall manager, or
the members will be given the exact information it needs to complete the
task required.